Add files via upload
This code is for classifying text data from 
the Spooky Authors (train)dataset, which can be downloaded from Kaggle
(https://www.kaggle.com/competitions/spooky-author-identification/data). 
The dataset is used to train four machine learning models, including an 
MLP, LSTM, CNN, and RNN, to classify the text into three categories(authors).
The code preprocesses the text data by vectorizing it using Tf-idf, padding
the sequences to the same length, and encoding the labels. The models are then 
trained and their performance is visualized using plots of the mean squared error
and accuracy during training, as well as ROC curves for each model.